Add files via upload
Recently, the US Government published the National Cybersecurity Strategy (National-Cybersecurity-Strategy-2023.pdf (whitehouse.gov))Links to an external site. with several pillars, including Defending Critical Infrastructure. 

In this assignment you are an analyst working for a company in a critical infrastructure sector of your choosing (https://www.cisa.gov/topics/critical-infrastructure-security-and-resilience/critical-infrastructure-sectorsLinks to an external site.). You have been asked by your leadership to review the recently published National Cybersecurity Strategy’s pillar on defending critical infrastructure and write a memorandum summing up the strategy’s pillar and its impact on the critical infrastructure sector of your choosing. The goal of the memorandum is to help your company develop an organizational strategy that aligns with the ones published by the US Government.