Add files via upload
Imagine you are a US cyber security official engaging in a training exercise. Your supervisor provides you with a case study scenario 

Your supervisor asks you to submit a document that describes, with respect to the scenario, a desired strategic end-state, strategic tenets or pillars (perhaps 3-5) to support that end-state, and a literature review to justify your strategic decisions. This literature review must include an analysis of course materials, and it can also include other materials that you find reliable, useful, and relevant.